moved minimum of doubles to utils/math.c and made gnsf_neville static
